backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform  # optional if you use the client lib
